server/requirements.txt
amqp==5.0.6
aniso8601==9.0.1
bidict==0.21.2
billiard==3.6.4.0
cachelib==0.1.1
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coolname==1.1.0
cryptography==3.4.7
dnspython==1.16.0
eventlet==0.30.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
Flask-Session==0.3.2
Flask-SocketIO==5.0.1
greenlet==1.0.0
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
kombu==5.0.2
MarkupSafe==1.1.1
mongo==0.2.0
numpy==1.20.2
prometheus-async==19.2.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
psutil==5.8.0
pycparser==2.20
pymongo==3.11.4
pyOpenSSL==20.0.1
python-engineio==4.0.1
python-socketio==5.1.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
rq==1.8.0
six==1.15.0
socketIO-client==0.5.7.2
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.59.0
Werkzeug==1.0.1
wrapt==1.12.1
